app/pyproject.toml
[tool.poetry]
name = "chatgpt-app"
version = "0.1.0"
description = ""
authors = ["Naohiro Otake <n_otake@openwork.co.jp>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
streamlit = "^1.20.0"
openai = "^0.27.2"
google-cloud-bigquery = "^3.9.0"
google-cloud-texttospeech = "^2.14.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
